fix: ensure UriFactory works under PHP 8.1 when provided with a relative URI
A call to `$uri->getScheme()` can return a `null`.
The `UriFactory::factory()` method calls that method on the generated URI to determine the scheme, if any, in order to provide a more specific URI instance.
To do so, it passes the return value to `strtolower()`.
This raises a deprecation in PHP 8.1, as passing `null` to that function is now deprecated.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>